Second refactor: chunking down and put functions to their relevant space
